% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate_assessment.R
\name{print.ate_assessment}
\alias{print.ate_assessment}
\title{Print method for ate_assessment objects}
\usage{
\method{print}{ate_assessment}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of class \code{ate_assessment}.}

\item{digits}{Number of decimal places to display (default: 2).}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Displays a concise summary of the performance assessment.
}
\examples{
assess <- ate_assessment(bias = 1.5, cv = 2.5, tea = 10)
print(assess)

}
