% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate_methods.R
\name{print.ate_specs}
\alias{print.ate_specs}
\title{Print method for ate_specs objects}
\usage{
\method{print}{ate_specs}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of class \code{ate_specs}.}

\item{digits}{Number of decimal places to display (default: 2).}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Displays a concise summary of allowable total error specifications
calculated from biological variation.
}
\examples{
ate <- ate_from_bv(cvi = 5.6, cvg = 7.5)
print(ate)

}
