% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_profile_methods.R
\name{print.precision_profile}
\alias{print.precision_profile}
\title{Print method for precision_profile objects}
\usage{
\method{print}{precision_profile}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{precision_profile}.}

\item{digits}{Number of significant digits to display (default: 3).}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Displays a concise summary of precision profile results, including model
fit and functional sensitivity estimates.
}
\examples{
# See ?precision_profile for examples

}
