% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_metric.R
\name{print.sigma_metric}
\alias{print.sigma_metric}
\title{Print method for sigma_metric objects}
\usage{
\method{print}{sigma_metric}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of class \code{sigma_metric}.}

\item{digits}{Number of decimal places to display (default: 2).}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Displays a concise summary of the sigma metric calculation.
}
\examples{
sm <- sigma_metric(bias = 1.5, cv = 2.0, tea = 10)
print(sm)

}
