% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_methods.R
\name{print.summary.precision_study}
\alias{print.summary.precision_study}
\title{Print method for summary.precision_study objects}
\usage{
\method{print}{summary.precision_study}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{summary.precision_study}.}

\item{digits}{Number of significant digits to display (default: 4).}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for summary.precision_study objects
}
