% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_profile_methods.R
\name{summary.precision_profile}
\alias{summary.precision_profile}
\title{Summary method for precision_profile objects}
\usage{
\method{summary}{precision_profile}(object, ...)
}
\arguments{
\item{object}{An object of class \code{precision_profile}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
An object of class \code{summary.precision_profile} containing summary
statistics.
}
\description{
Provides a detailed summary of precision profile results, including fitted
values, residuals, fit statistics, and functional sensitivity estimates.
}
\examples{
# See ?precision_profile for examples

}
