% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valytics-package.R
\docType{package}
\name{valytics-package}
\alias{valytics}
\alias{valytics-package}
\title{valytics: Statistical Methods for Analytical Method Comparison and Validation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides statistical methods for analytical method comparison and validation studies. Implements Bland-Altman analysis for assessing agreement between measurement methods (Bland & Altman (1986) \doi{10.1016/S0140-6736(86)90837-8}), Passing-Bablok regression for non-parametric method comparison (Passing & Bablok (1983) \doi{10.1515/cclm.1983.21.11.709}), and Deming regression accounting for measurement error in both variables (Linnet (1993) \doi{10.1093/clinchem/39.3.424}). Also includes tools for setting quality goals based on biological variation (Fraser & Petersen (1993) \doi{10.1093/clinchem/39.7.1447}) and calculating Six Sigma metrics. Commonly used in clinical laboratory method validation. Provides publication-ready plots and comprehensive statistical summaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcellogr/valytics}
  \item Report bugs at \url{https://github.com/marcellogr/valytics/issues}
}

}
\author{
\strong{Maintainer}: Marcello Grassi \email{marcello.grassi@tum.de} (\href{https://orcid.org/0009-0008-8904-4988}{ORCID})

}
\keyword{internal}
