% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valytics-package.R
\docType{package}
\name{valytics-package}
\alias{valytics}
\alias{valytics-package}
\title{valytics: Statistical Methods for Analytical Method Comparison and Validation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Statistical methods for analytical method comparison and validation studies. Implements Bland-Altman analysis, Passing-Bablok regression, and Deming regression for method comparison, along with tools for setting quality goals based on biological variation and calculating Six Sigma metrics. Commonly used in clinical laboratory method validation following CLSI guidelines. Provides publication-ready plots and comprehensive statistical summaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcellogr/valytics}
  \item Report bugs at \url{https://github.com/marcellogr/valytics/issues}
}

}
\author{
\strong{Maintainer}: Marcello Grassi \email{marcello.grassi@tum.de} (\href{https://orcid.org/0009-0008-8904-4988}{ORCID})

}
\keyword{internal}
