% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valytics-package.R
\docType{package}
\name{valytics-package}
\alias{valytics}
\alias{valytics-package}
\title{valytics: Statistical Methods for Analytical Method Comparison and Validation}
\description{
Statistical methods for analytical method comparison and validation studies. Implements approaches commonly used in clinical laboratory method validation.
}
\author{
\strong{Maintainer}: Marcello Grassi \email{marcello.grassi@tum.de} (\href{https://orcid.org/0009-0008-8904-4988}{ORCID})

}
\keyword{internal}
